+++ tests/Quotinator.Core.Tests/Data/SqliteQuoteServiceConversationTests.cs
@@ -8,7 +8,6 @@ using Quotinator.Data.Repositories;
 using Quotinator.Data.Testing.NoOps;
 using Quotinator.Engine.Database;
 using Quotinator.Engine.Entities;
-using Quotinator.Engine.Repositories;
 using Quotinator.Engine.Services;
 
 namespace Quotinator.Core.Tests.Data;