+++ addon/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-07-11 18:16 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-07-11 19:05 UTC] — do not edit by hand.*
 
 # Changelog